requirements.txt
astroid==2.2.5
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.3
django-braces==1.13.0
djangorestframework==3.9.4
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.2
wrapt==1.11.2
